pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = [
    "api_adapters", "api_models", "api_runtime",
    "cli_review", "config", "generate_report_pdf",
    "graph", "llm_factory", "run_pipeline", "server", "state",
]
packages = ["agents", "auth", "evals", "tools", "tests"]

[project]
name = "research-console"
version = "0.2.0"
description = "LangGraph research pipeline with human-in-the-loop and browser UI"
requires-python = ">=3.9"
readme = "README.md"

dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agents", "auth", "tools", "evals", "config", "state", "graph", "llm_factory"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_return_any = false
warn_unused_ignores = false
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
pythonpath = ["."]
